face_recognition_service/requirements.txt
numpy>=1.24
opencv-python>=4.8
Pillow>=10.0
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.5
matplotlib>=3.7
